moves the fn functions files into the svgTools files since they do the same thing and it could be confusing
Took 1 minute
